% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labels.R
\name{get_labels}
\alias{get_labels}
\title{Get all Labels}
\usage{
get_labels(catalog)
}
\arguments{
\item{catalog}{A list retured by read_catalog()}
}
\value{
A vector of unique labels in a catalog
}
\description{
Get all labels from a given catalog
}
