% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Get all values in a catalog}
\usage{
get_values(catalog, return_df = FALSE)
}
\arguments{
\item{catalog}{A list returned by read_catalog}

\item{return_df}{A boolean indicating if a dataframe should be returned}
}
\value{
A list or a dataframe of all values in the catalog
}
\description{
Get all values in a catalog in either a list of a dataframe
}
