% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogr.R
\name{initialize}
\alias{initialize}
\title{Initialize AWS configuration for the library.}
\usage{
initialize(bucket = NULL, profile = "default")
}
\arguments{
\item{profile}{}
}
\description{
Initialize AWS configuration for the library.
}
