% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_catalog.R
\name{read_catalog}
\alias{read_catalog}
\title{Read a Catalog}
\usage{
read_catalog(yaml_filepath = file.path(getwd(), "catalogr.yaml"))
}
\arguments{
\item{yaml_filepath}{the filepath of the saved catalog file}
}
\description{
A convenient yaml wrapper to read a catalog yaml file that is saved through `caralogr()`
}
