% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogr.R
\name{read_dataset}
\alias{read_dataset}
\title{Read a named dataset from the data catalogue bucket.}
\usage{
read_dataset(dataset_name)
}
\arguments{
\item{dataset_name}{}
}
\value{
data frame of the dataset
}
\description{
Read a named dataset from the data catalogue bucket.
}
