% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_catalog}
\alias{search_catalog}
\title{Search Catalog}
\usage{
search_catalog(catalog, search_value, label = NULL)
}
\arguments{
\item{catalog}{A list retured by read_catalog}

\item{search_value}{a search string}

\item{label}{A string indicating the label we wish to search. If NULL it matches all labels.}
}
\value{
a list of matched filenames
}
\description{
Search the catalog and return a filename.
}
