% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_catalog.R
\name{search_label}
\alias{search_label}
\title{Search a label in a catalog}
\usage{
search_label(catalog, label_str)
}
\arguments{
\item{catalog}{A list returned by read_catalog}

\item{label_str}{A label to search}
}
\value{
A catalog with only the label in label_str
}
\description{
Get a catalog for a given label
}
