% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogr.R
\name{catalogr}
\alias{catalogr}
\title{}
\usage{
catalogr(dir = getwd(), fileformat = list(.R = "#@"))
}
\arguments{
\item{dir}{current working directory}

\item{fileformat}{list with the keys as file type and value with the token}
}
\description{

}
